+ DELETE

+++ twake/backend/node/test/e2e/workspaces/workspace-users.spec.ts
@@ -1,6 +1,6 @@
 import { afterAll, beforeAll, describe, expect, it } from "@jest/globals";
 import { init, TestPlatform } from "../setup";
-import { TestDbService } from "../utils.prepare.db";
+import { TestDbService, uuid } from "../utils.prepare.db";
 import { v4 as uuidv4 } from "uuid";
 
 describe("The /workspace users API", () => {
@@ -41,21 +41,27 @@ describe("The /workspace users API", () => {
     platform = await init({
       services: ["database", "pubsub", "webserver", "user", "workspaces", "auth"],
     });
-    await platform.database.getConnector().drop();
+
+    if ((platform.database as any).type == "mongodb") {
+      await platform.database.getConnector().drop();
+    }
     await platform.database.getConnector().init();
     testDbService = new TestDbService(platform);
     await testDbService.createCompany(companyId);
     const ws0pk = { id: uuidv4(), group_id: companyId };
     const ws1pk = { id: uuidv4(), group_id: companyId };
     const ws2pk = { id: uuidv4(), group_id: companyId };
+    const ws3pk = { id: uuidv4(), group_id: companyId };
     await testDbService.createWorkspace(ws0pk);
     await testDbService.createWorkspace(ws1pk);
     await testDbService.createWorkspace(ws2pk);
+    await testDbService.createWorkspace(ws3pk);
     await testDbService.createUser([ws0pk, ws1pk]);
     await testDbService.createUser([ws2pk], "admin");
     await testDbService.createUser([ws2pk], undefined, "admin");
     await testDbService.createUser([ws2pk], undefined, "member");
     await testDbService.createUser([], "guest");
+    await testDbService.createUser([ws3pk], "guest", "member");
     ends();
   });
 
@@ -300,7 +306,7 @@ describe("The /workspace users API", () => {
     it("should 404 when user not found in workspace", async done => {
       const workspaceId = testDbService.workspaces[2].workspace.id;
       const userId = testDbService.workspaces[2].users[1].id;
-      const anotherWorkspaceUserId = testDbService.workspaces[0].users[0].id;
+      const anotherWorkspaceUserId = testDbService.workspaces[3].users[0].id;
       const jwtToken = await platform.auth.getJWTToken({ sub: userId });
 
       const response = await platform.app.inject({
@@ -374,6 +380,24 @@ describe("The /workspace users API", () => {
       done();
     });
 
+    it("should 404 when user not found in workspace", async done => {
+      const companyId = testDbService.company.id;
+      const workspaceId = testDbService.workspaces[2].workspace.id;
+      const userId = testDbService.workspaces[2].users[1].id;
+
+      const jwtToken = await platform.auth.getJWTToken({ sub: userId });
+      const response = await platform.app.inject({
+        method: "DELETE",
+        url: `${url}/companies/${companyId}/workspaces/${workspaceId}/users/${nonExistentId}`,
+        headers: { authorization: `Bearer ${jwtToken}` },
+      });
+
+      console.log(response.json());
+      expect(response.statusCode).toBe(404);
+
+      done();
+    });
+
     it("should 200 when ok", async done => {
       const companyId = testDbService.company.id;
       const workspaceId = testDbService.workspaces[2].workspace.id;
@@ -381,19 +405,29 @@ describe("The /workspace users API", () => {
       const anotherUserId = testDbService.workspaces[2].users[2].id;
 
       const jwtToken = await platform.auth.getJWTToken({ sub: userId });
-      const response = await platform.app.inject({
+      let response = await platform.app.inject({
         method: "DELETE",
         url: `${url}/companies/${companyId}/workspaces/${workspaceId}/users/${anotherUserId}`,
         headers: { authorization: `Bearer ${jwtToken}` },
       });
 
-      expect(response.statusCode).toBe(501);
+      expect(response.statusCode).toBe(204);
+
+      response = await platform.app.inject({
+        method: "GET",
+        url: `${url}/companies/${companyId}/workspaces/${workspaceId}/users`,
+        headers: { authorization: `Bearer ${jwtToken}` },
+      });
+
+      expect(response.statusCode).toBe(200);
+      const resources = response.json()["resources"];
+      expect(resources.find((a: { user_id: uuid }) => a.user_id === anotherUserId)).toBeUndefined();
 
       done();
     });
   });
 
-  describe("The POST /workspaces/users/invite route", () => {
+  describe.skip("The POST /workspaces/users/invite route", () => {
     it("should 401 when not authenticated", async done => {
       const companyId = testDbService.company.id;
       const userId = testDbService.users[0].id;
